remove msvc warnings

+++ h/registers.h
@@ -8,4 +8,4 @@ extern const char *float_reg_abi_map[];
 size_t get_register_id(const char *);
 size_t get_float_register_id(const char *);
 
-size_t get_immediate(const char *, size_t *);
+int get_immediate(const char *, size_t *);
